requirements.txt
gradio==5.29.0
openai==0.28.0
python-dotenv==1.1.0
